separate staffgroup layout from staffgroup element
